unit test gym-csle-stopping-game

+++ simulation-system/libs/gym-csle-stopping-game/src/gym_csle_stopping_game/util/stopping_game_util.py
@@ -15,13 +15,13 @@ class StoppingGameUtil:
     """
 
     @staticmethod
-    def b1() -> npt.NDArray[np.int_]:
+    def b1() -> npt.NDArray[np.float_]:
         """
         Gets the initial belief
 
         :return: the initial belief
         """
-        return np.array([1, 0, 0])
+        return np.array([1.0, 0.0, 0.0])
 
     @staticmethod
     def state_space():
